Change keyword_method_names test
This test is just meant to capture the fact that we're already deviating
with what the Ruby parser would do. We're fine with these changes

The only potentially concerning thing is that to get the test to
somewhat still match up withh the old behavior, we manually added an
extra `end` keyword. The justification for why that's like that is that
it's a separate ticket to track mismatched `end` keywords (for def,
begin, etc.) that we'll get to eventually (maybe even shortly).